VBoxManage,manual: Use the new RTCrX509Certificate_WriteToVfsFile function to write out the signing certificate. [nits] bugref:9699
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@84167 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Frontends/VBoxManage/VBoxManageAppliance.cpp
@@ -2002,12 +2002,12 @@ static int doTheOvaSigning(PRTCRX509CERTIFICATE pCertificate, RTCRKEY hPrivateKe
                         rc = RTVfsMemFileCreate(NIL_RTVFSIOSTREAM, _8K, &hVfsFileSignature);
                         if (RT_SUCCESS(rc))
                         {
-                            rc = RTVfsFilePrintf(hVfsFileSignature, "%s(%s) = %.*Rhxs\n\n",
-                                                 pszDigestType, pszManifestName, cbSignature, pvSignature);
+                            rc = (int)RTVfsFilePrintf(hVfsFileSignature, "%s(%s) = %.*Rhxs\n\n",
+                                                      pszDigestType, pszManifestName, cbSignature, pvSignature);
                             if (RT_SUCCESS(rc))
                             {
-                                rc = RTCrX509Certificate_WriteToVfsFile(hVfsFileSignature, pCertificate,
-                                                                        RTErrInfoInitStatic(pErrInfo));
+                                rc = (int)RTCrX509Certificate_WriteToVfsFile(hVfsFileSignature, pCertificate,
+                                                                             RTErrInfoInitStatic(pErrInfo));
                                 if (RT_SUCCESS(rc))
                                 {
                                     if (fPkcs7)
